refactor: extract buildApp into app/ modules, decompose failover-loop, and improve stream-oa2ant
- Split buildApp() (575 lines) into 4 focused modules in src/app/:
  compose-container.ts, create-app.ts, register-hooks.ts, register-routes.ts
- Extract failover-loop into 3 sub-modules:
  iteration-setup.ts, reject-helpers.ts, resilience-processor.ts
- Simplify stream-oa2ant.ts stream transform logic
- Remove unused hook-registry.ts, add emit_error to pipeline types
- Add harness design docs and unit tests for new modules
