+++ .bg1-result.md
@@ -1,0 +1,42 @@
+# BG1 Result: buildApp() 拆分为 4 个函数
+
+## 变更概要
+
+将 `router/src/index.ts` 中 575 行的 `buildApp()` 拆分为 4 个独立函数，放入 `router/src/app/` 目录。
+
+## 新增文件
+
+| 文件 | 函数 | 职责 |
+|------|------|------|
+| `router/src/app/create-app.ts` | `createAppInstance(opts)` | 创建 Fastify 实例 + 全局 hooks（errorHandler、onSend 信封包装、EPIPE 防护） |
+| `router/src/app/compose-container.ts` | `composeContainer(db, opts, app)` | 注册所有服务到 ServiceContainer（matcher、semaphore、tracker、plugins、formats 等） |
+| `router/src/app/register-hooks.ts` | `registerAppHooks(app, db, container, deps)` | 注册 auth middleware、内置 hooks、proxy handlers，构建 StateRegistry |
+| `router/src/app/register-routes.ts` | `registerRoutes(app, opts)` | 注册 admin routes、静态文件、/health、定时任务，组装 close 函数 |
+
+## 修改文件
+
+| 文件 | 变更 |
+|------|------|
+| `router/src/index.ts` | buildApp 从 575 行大函数变为 ~50 行编排函数；移除已迁移到 app/ 的 import；保留 `initializeProviderState`、`main`、所有 export 签名不变 |
+| `router/tests/create-app.test.ts` | 新增：验证 buildApp 返回结构和 /health 端点（2 个测试用例） |
+
+## 关键路径修正
+
+由于文件从 `src/` 移至 `src/app/`，`__dirname` 深了一层，以下路径需要修正：
+
+- `loadRecommendedConfig` / `startUpgradeChecker`: `../config` → `../../config`
+- `pluginRegistry.scanPluginsDir`: `../plugins/transform` → `../../plugins/transform`
+- `frontendDist` 静态文件: `../frontend-dist` → `../../frontend-dist`
+
+## 验证结果
+
+| 检查项 | 结果 |
+|--------|------|
+| `npx tsc --noEmit` | 0 error |
+| `npm run lint` | 0 error, 0 warning |
+| `npm test` | 145 files passed, 1787 tests passed, 5 skipped |
+| `tests/create-app.test.ts` | 2/2 passed |
+
+## 行为变更
+
+无。纯提取重构，所有 export 签名、请求处理流程、错误处理逻辑保持不变。
