+++ .bg2-result.md
@@ -1,0 +1,48 @@
+# BG2 结果：合并 hook-registry 到 pipeline.ts + emit 异常降级
+
+## 变更摘要
+
+| 操作 | 文件 |
+|------|------|
+| **删除** | `router/src/proxy/pipeline/hook-registry.ts` |
+| **重写** | `router/src/proxy/pipeline/pipeline.ts` — 合并 HookRegistry 功能 + emit 异常降级 |
+| **修改** | `router/src/proxy/pipeline/types.ts` — PipelineHook 增加 `core?: boolean` |
+| **修改** | `router/src/proxy/pipeline/register-hooks.ts` — 删除 hookRegistry 引用 |
+| **修改** | `router/src/app/register-hooks.ts` — hookRegistry.getAll() → proxyPipeline.getAllHooks() |
+| **新增** | `tests/unit/pipeline-emit-error-handling.test.ts` — 3 个测试用例 |
+| **修改** | `router/tests/proxy/pipeline/pipeline.test.ts` — 适配新的 emit 降级行为 |
+
+## 关键设计
+
+### emit 异常降级
+
+```typescript
+async emit(phase, ctx) {
+  for (const hook of hooks) {
+    if (hook.core === true) {
+      await hook.execute(ctx);  // 异常传播，中断后续 hook
+    } else {
+      try {
+        await hook.execute(ctx);
+      } catch (err) {
+        ctx.request.log.warn({ hook: hook.name, err }, "non-core hook error");
+        // 继续执行后续 hook
+      }
+    }
+  }
+}
+```
+
+### HookSummary 迁移
+
+`HookSummary` 接口和 `getByPhase()`/`getAllHooks()` 方法从 HookRegistry 合并到 ProxyPipeline，Admin API 通过 `proxyPipeline.getAllHooks()` 获取数据。
+
+## 验证结果
+
+- tsc --noEmit: 0 error
+- eslint（源码文件）: 0 error 0 warning
+- vitest（pipeline 相关）: 33 tests passed
+- vitest（全量）: 1851 passed / 3 failed（均为既有失败，非本次引入）
+  - `stream-oa2ant.test.ts` TC-3-03: 既有问题
+  - `transform-rules.test.ts` reload: 既有问题
+  - `useProviderForm.test.ts`: 缺少源文件，既有问题
