+++ .bg3-result.md
@@ -1,0 +1,35 @@
+# BG3 结果：stream-oa2ant 映射表重写
+
+## 变更文件
+
+| 文件 | 操作 | 行数变化 |
+|------|------|---------|
+| `router/src/proxy/transform/stream-oa2ant.ts` | 重写 | 223 → 192（-14%） |
+| `tests/unit/stream-oa2ant.test.ts` | 新增 | 127 行（7 个测试用例） |
+
+## 重构方法
+
+核心改动：将分散在 `ensureBlockState`、`handleToolCallDelta`、`closeCurrentBlock` 三处的 "close current block → increment index → emit block_start" 重复模式，统一到 `transitionBlock()` 方法。
+
+| 原方法 | 重构后 |
+|--------|--------|
+| `ensureBlockState(target)` | `transitionBlock(target, blockContent)` — 统一 block 切换 |
+| `handleToolCallDelta` 内的 block 管理 | `openToolBlock(idx, id, name)` — 每个 tool call 独立创建 block |
+| `closeCurrentBlock` | 保留，委托给 `emitBlockStop()` |
+| 内联 `pushAnthropicSSE` 调用 | 提取 `emitDelta()`、`emitJsonDelta()`、`emitBlockStop()` 减少重复 |
+| 内联 usage 提取 | 保留在 `processEvent` 内（仅调用一处） |
+| `ensureMessageStart` 条件检查 | 独立方法（与原代码一致） |
+
+关键设计决策：
+- `transitionBlock` 对 text/thinking 状态做 same-state 短路（避免重复 block_start）
+- `openToolBlock` 不走 `transitionBlock`，因为 tool_use 需要每个 call 独立 block（即使状态已是 tool_use）
+- `toolCallBufferedArgs`（原 `toolCallBlocks`）保留，使用 `Map<number, string>` 简化结构（原代码存储了未使用的 id/name 字段）
+
+## 验证结果
+
+```
+✓ tests/unit/stream-oa2ant.test.ts — 7/7 passed
+✓ router/tests/proxy/transform/stream-transform-oa2ant.test.ts — 6/6 passed
+✓ tsc --noEmit — 0 errors
+✓ eslint — 0 warnings
+```
